Make Button_RightTapped and LoadShellMenuItems common to all widgets
