+++ src/Files.App/UserControls/Widgets/WidgetCardItem.cs
@@ -10,5 +10,7 @@ namespace Files.App.UserControls.Widgets
 	public abstract class WidgetCardItem : ObservableObject
 	{
 		public string Path;
+
+		public virtual object Item { get; set; }
 	}
 }
